web-platform/requirements.txt
amqp==5.2.0
asgiref==3.8.1
async-timeout==5.0.1
asyncpg==0.29.0
attrs==24.2.0
autobahn==24.4.2
Automat==24.8.1
axios==0.4.0
billiard==4.2.1
boto3==1.35.72
botocore==1.35.72
celery==5.4.0
certifi==2024.8.30
cffi==1.17.1
channels==4.1.0
channels-redis==4.2.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
constantly==23.10.4
cryptography==43.0.1
daphne==4.1.2
Django==5.1.1
django-appconf==1.0.6
django-celery-progress==1.0.5
django-celery-results==2.5.1
django-cors-headers==4.4.0
django-environ==0.12.0
django-imagekit==5.0.0
django-redis==5.4.0
django-resumable==0.2.0
django-storages==1.14.4
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
jmespath==1.0.1
kombu==5.4.2
lxml==5.3.0
markdown-it-py==3.0.0
mdurl==0.1.2
msgpack==1.1.0
numpy==2.2.4
pilkit==3.0
pillow==10.4.0
prompt_toolkit==3.0.48
psycopg2-binary==2.9.9
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
Pygments==2.18.0
PyJWT==2.9.0
pyOpenSSL==24.2.1
python-dateutil==2.9.0.post0
python-magic==0.4.27
PyYAML==6.0.2
redis==5.1.1
requests==2.32.3
rich==13.9.1
s3transfer==0.10.4
service-identity==24.1.0
six==1.16.0
sqlparse==0.5.1
tomli==2.2.1
Twisted==24.7.0
txaio==23.1.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
vine==5.1.0
wcwidth==0.2.13
zope.interface==7.1.0
